Refactor: bases_* 폴더에 있는 코드를 srcs에 폴더별로 분류함.
What: base_headers 와 bases_srcs 에 있는 코드들을 srcs에 옮기면서 하나하나씩 해결해나가면 좋을 것 같음
Why: 소스코드에서 header include 할 때, 여러 폴더에서 참조하는 것이 불편하고, 에러를 발생시킬 요지가 많아서
