--- base_headers/all_header.hpp
@@ -1,22 +1,0 @@
-#pragma once
-
-#include <iostream>
-using namespace std;
-	
-// #include "./Buffer.hpp"
-// #include "./Cgi.hpp"
-// #include "./Cgi.hpp"
-// #include "./Client.hpp"
-#include "./Config.hpp"
-// #include "./ConfigWebserver.hpp"
-// #include "./FdSet.hpp"
-// #include "./Http.hpp"
-// #include "./Method.hpp"
-#include "./Path.hpp"
-#include "./Server.hpp"
-// #include "./Socket.hpp"
-// #include "./Url.hpp"
-#include "./utils.hpp"
-// #include "./UtilsTemplate.hpp"
-// #include "./base_headers/Webserver.hpp"
-
