--- base_headers/utils.hpp
@@ -1,8 +1,0 @@
-#pragma once
-#include <string>
-#include <unistd.h>
-#include <iostream>
-#include <vector>
-
-void ft_split_vector(std::vector<std::string> &vec, const std::string& str, const char *delim);
-
