--- yunslee_test/server_parsing/parsing1
@@ -1,14 +1,0 @@
-server
-{
-{
-		
-}
-	
-}
-server
-{
-	server_name 127.0.0.1
-	port 80
-	error_page error.html
-	index index2.html
-}
